feature(authentication): added routing and auth components
