+++ src/authComponents/Profile.js
@@ -1,0 +1,3 @@
+export default function Profile() {
+    
+}
