+++ src/authComponents/Register.js
@@ -1,0 +1,58 @@
+import { useState } from 'react';
+import { Link } from 'react-router-dom';
+import './forms.css';
+
+function Register() {
+
+    const [email, setEmail] = useState('');
+    const [password, setPassword] = useState('')
+    const [confirmPassword, setConfirmPassword] = useState('')
+    const [error, setError] = useState('')
+
+    return (
+        <div className='center'>
+            <div className='auth'>
+                <h1>Register</h1>
+                {error && <div className='auth__error'>{error}</div>}
+
+                <form name='registration_form'>
+                    <input
+                       type='email'
+                       value={email}
+                       placeholder='Enter your email'
+                       required
+                       onChange={e => setEmail(e.target.value)} 
+                    />
+
+                    <input
+                       type='password'
+                       value={password}
+                       required
+                       placeholder='Enter your password'
+                       onChange={e => setPassword(e.target.value)} 
+                    />
+
+                    <input
+                       type='password'
+                       value={confirmPassword}
+                       required
+                       placeholder='Confirm password'
+                       onChange={e => setConfirmPassword(e.target.value)} 
+                    />
+
+                    <button type='submit'>Register</button>
+                </form>
+                <span>
+                    Already have an account?
+                    <Link to='/login'>login</Link>
+
+                    Or Log in as guest
+                </span>
+
+            </div>
+
+        </div>
+    )
+}
+
+export default Register;
