+++ src/authComponents/VerifyEmail.js
@@ -1,0 +1,3 @@
+export default function VerifyEmail() {
+    
+}
